requirements.txt
aiohttp==3.9.5
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
bcrypt==4.1.2
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.7
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cache2==0.2.1
fastapi-cli==0.0.4
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
flower==2.0.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanize==4.9.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.3.7
makefun==1.15.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
orjson==3.10.3
packaging==24.0
pendulum==3.0.0
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.46
pwdlib==0.2.0
pycparser==2.22
pydantic==2.7.2
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.3
Pygments==2.18.0
PyJWT==2.8.0
pytest==8.2.2
pytest-asyncio==0.23.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==4.6.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
time-machine==2.14.1
tornado==6.4
typer==0.12.3
typing_extensions==4.12.1
tzdata==2024.1
ujson==5.10.0
uvicorn==0.30.1
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
gunicorn
